@@ -6,69 +6,67 @@
     "start": "ng serve",
     "build": "ng build",
     "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "lint": "ng lint"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.2.16",
-    "@angular/cdk": "^12.2.13",
-    "@angular/common": "~12.2.16",
-    "@angular/compiler": "~12.2.16",
-    "@angular/core": "~12.2.16",
-    "@angular/forms": "~12.2.16",
-    "@angular/localize": "^12.2.16",
-    "@angular/material": "^12.2.13",
-    "@angular/platform-browser": "~12.2.16",
-    "@angular/platform-browser-dynamic": "~12.2.16",
-    "@angular/router": "~12.2.16",
-    "@ng-bootstrap/ng-bootstrap": "~10",
-    "angular-google-charts": "~2",
-    "angular-plotly.js": "^4.0.4",
-    "angular-resizable-element": "^3.3.3",
-    "bootstrap": "^4.4.1",
-    "chart.js": "^3.9.1",
-    "datatables.net": "^1.10.20",
-    "datatables.net-dt": "^1.10.20",
-    "datatables.net-select": "^1.3.1",
-    "datatables.net-select-dt": "^1.3.1",
+    "@angular/animations": "~19.2.14",
+    "@angular/cdk": "^19.2.18",
+    "@angular/common": "~19.2.14",
+    "@angular/compiler": "~19.2.14",
+    "@angular/core": "~19.2.14",
+    "@angular/forms": "~19.2.14",
+    "@angular/localize": "^19.2.14",
+    "@angular/material": "^19.2.18",
+    "@angular/platform-browser": "~19.2.14",
+    "@angular/platform-browser-dynamic": "~19.2.14",
+    "@angular/router": "~19.2.14",
+    "@ng-bootstrap/ng-bootstrap": "^18.0.0",
+    "@popperjs/core": "^2.11.8",
+    "angular-google-charts": "^16.0.2",
+    "angular-plotly.js": "^6.0.0",
+    "angular-resizable-element": "^7.0.2",
+    "bootstrap": "^5.3.6",
+    "chart.js": "4.5",
+    "datatables.net": "^1.12.1",
+    "datatables.net-dt": "^1.12.1",
+    "datatables.net-select": "^1.4.0",
+    "datatables.net-select-dt": "^1.4.0",
     "font-awesome": "^4.7.0",
     "jquery": "^3.4.1",
     "material-design-icons": "^3.0.1",
-    "ng-dynamic-component": "~9",
-    "ng-multiselect-dropdown": "^0.3.8",
-    "ng2-charts": "^3.1.2",
-    "plotly.js-dist-min": "^3.0.1",
-    "popper.js": "^1.16.1",
-    "rxjs": "~6.6.3",
-    "rxjs-compat": "^6.5.4",
+    "ng-dynamic-component": "^10.8.2",
+    "ng-multiselect-dropdown": "^1.0.0",
+    "ng2-charts": "^6.0.1",
+    "plotly.js-dist-min": "^3.1.0",
+    "rxjs": "^7.8.1",
+    "rxjs-compat": "^6.6.7",
     "tslib": "^2.4.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.15.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^12.2.17",
-    "@angular/cli": "^12.2.17",
-    "@angular/compiler-cli": "~12.2.16",
-    "@angular/language-service": "~12.2.16",
-    "@types/datatables.net": "^1.10.18",
-    "@types/datatables.net-select": "^1.2.6",
+    "@angular-devkit/build-angular": "^19.2.14",
+    "@angular-eslint/builder": "^20.0.0",
+    "@angular-eslint/eslint-plugin": "^20.0.0",
+    "@angular-eslint/eslint-plugin-template": "^20.0.0",
+    "@angular-eslint/schematics": "^20.0.0",
+    "@angular-eslint/template-parser": "^20.0.0",
+    "@angular/cli": "^19.2.14",
+    "@angular/compiler-cli": "~19.2.14",
+    "@angular/language-service": "~19.2.14",
     "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "~2.0.3",
     "@types/jquery": "^3.3.32",
     "@types/lodash": "^4.14.149",
     "@types/node": "^12.11.1",
     "@types/plotly.js-dist-min": "^2.3.4",
+    "@typescript-eslint/eslint-plugin": "^7.2.0",
+    "@typescript-eslint/parser": "^7.2.0",
+    "eslint": "^8.57.0",
     "jasmine-core": ">=3.8.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.3.4",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^12.2.7",
-    "protractor": "~7.0.0",
+    "ng-packagr": "^19.2.2",
     "ts-node": "~7.0.0",
-    "typescript": "^4.3.5"
+    "typescript": "^5.8.3"
   },
   "overrides": {
     "@types/eslint": "8.4.3"
