@@ -3,7 +3,7 @@
  * API for Ig Receptor gene data
  *
  * OpenAPI spec version: 1.0
- * 
+ *
  *
  * NOTE: This class is auto generated by the swagger code generator program.
  * https://github.com/swagger-api/swagger-codegen.git
@@ -12,11 +12,10 @@
 /* tslint:disable:no-unused-variable member-ordering */
 
 import { Inject, Injectable, Optional }                      from '@angular/core';
-import { HttpClient, HttpHeaders, HttpParams,
-         HttpResponse, HttpEvent }                           from '@angular/common/http';
+import { HttpClient, HttpHeaders, HttpParams, HttpResponse, HttpEvent }                           from '@angular/common/http';
 import { CustomHttpUrlEncodingCodec }                        from '../encoder';
 
-import { Observable }                                        from 'rxjs/Observable';
+import { Observable }                                        from 'rxjs';
 
 
 import { BASE_PATH, COLLECTION_FORMATS }                     from '../variables';
@@ -57,9 +56,9 @@ export class GenomicService {
 
     /**
      * Returns information on all samples
-     * 
-     * @param species 
-     * @param dataset 
+     *
+     * @param species
+     * @param dataset
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -104,8 +103,8 @@ export class GenomicService {
 
     /**
      * Return genotypes for all subjects of the specified species in the specified data type
-     * 
-     * @param species 
+     *
+     * @param species
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -146,9 +145,9 @@ export class GenomicService {
 
     /**
      * Returns the list of annotated assemblies for the selected species and datasets
-     * 
-     * @param species 
-     * @param dataSets 
+     *
+     * @param species
+     * @param dataSets
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -193,8 +192,8 @@ export class GenomicService {
 
     /**
      * Returns the list of data sets for the selected species
-     * 
-     * @param species 
+     *
+     * @param species
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -235,9 +234,9 @@ export class GenomicService {
 
     /**
      * Returns information and statistics on the dataset
-     * 
-     * @param species 
-     * @param dataset 
+     *
+     * @param species
+     * @param dataset
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -282,11 +281,11 @@ export class GenomicService {
 
     /**
      * Returns the position of the first feature matching the specified string
-     * 
-     * @param species 
-     * @param dataset 
-     * @param refSeqName 
-     * @param featureString 
+     *
+     * @param species
+     * @param dataset
+     * @param refSeqName
+     * @param featureString
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -339,9 +338,9 @@ export class GenomicService {
 
     /**
      * Returns the inferred genotype (in MiAIRR format) of the specified patient
-     * 
-     * @param species 
-     * @param patientName 
+     *
+     * @param species
+     * @param patientName
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -386,14 +385,14 @@ export class GenomicService {
 
     /**
      * Returns nucleotide sequences from selected reference or multiple references (separate multiple reference names with &#39;,&#39;)
-     * 
-     * @param species 
-     * @param genomicDatasets 
-     * @param pageNumber 
-     * @param pageSize 
-     * @param filter 
-     * @param sortBy 
-     * @param cols 
+     *
+     * @param species
+     * @param genomicDatasets
+     * @param pageNumber
+     * @param pageSize
+     * @param filter
+     * @param sortBy
+     * @param cols
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -461,7 +460,7 @@ export class GenomicService {
 
     /**
      * Returns the list of species for which information is held
-     * 
+     *
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -498,10 +497,10 @@ export class GenomicService {
 
     /**
      * Returns information on the selected sample
-     * 
-     * @param species 
-     * @param dataset 
-     * @param sampleId 
+     *
+     * @param species
+     * @param dataset
+     * @param sampleId
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -550,14 +549,14 @@ export class GenomicService {
 
     /**
      * Returns a list of subjects in the selected datasets
-     * 
-     * @param species 
-     * @param genomicDatasets 
-     * @param pageNumber 
-     * @param pageSize 
-     * @param filter 
-     * @param sortBy 
-     * @param cols 
+     *
+     * @param species
+     * @param genomicDatasets
+     * @param pageNumber
+     * @param pageSize
+     * @param filter
+     * @param sortBy
+     * @param cols
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
