@@ -3,7 +3,7 @@
  * API for Ig Receptor gene data
  *
  * OpenAPI spec version: 1.0
- * 
+ *
  *
  * NOTE: This class is auto generated by the swagger code generator program.
  * https://github.com/swagger-api/swagger-codegen.git
@@ -16,7 +16,7 @@ import { HttpClient, HttpHeaders, HttpParams,
          HttpResponse, HttpEvent }                           from '@angular/common/http';
 import { CustomHttpUrlEncodingCodec }                        from '../encoder';
 
-import { Observable }                                        from 'rxjs/Observable';
+import { Observable }                                        from 'rxjs';
 
 
 import { BASE_PATH, COLLECTION_FORMATS }                     from '../variables';
@@ -57,10 +57,10 @@ export class RefbookService {
 
     /**
      * Returns sequences of all alleles in an ASC
-     * 
-     * @param species 
-     * @param chain 
-     * @param asc 
+     *
+     * @param species
+     * @param chain
+     * @param asc
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -109,9 +109,9 @@ export class RefbookService {
 
     /**
      * Returns the list of ASCs in a given chain for a given species
-     * 
-     * @param species 
-     * @param chain 
+     *
+     * @param species
+     * @param chain
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -156,10 +156,10 @@ export class RefbookService {
 
     /**
      * Returns data for the overview refbook component
-     * 
-     * @param species 
-     * @param chain 
-     * @param asc 
+     *
+     * @param species
+     * @param chain
+     * @param asc
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -208,7 +208,7 @@ export class RefbookService {
 
     /**
      * Returns the list of species and chains for which information is held
-     * 
+     *
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
