@@ -3,7 +3,7 @@
  * API for Ig Receptor gene data
  *
  * OpenAPI spec version: 1.0
- * 
+ *
  *
  * NOTE: This class is auto generated by the swagger code generator program.
  * https://github.com/swagger-api/swagger-codegen.git
@@ -12,11 +12,10 @@
 /* tslint:disable:no-unused-variable member-ordering */
 
 import { Inject, Injectable, Optional }                      from '@angular/core';
-import { HttpClient, HttpHeaders, HttpParams,
-         HttpResponse, HttpEvent }                           from '@angular/common/http';
+import { HttpClient, HttpHeaders, HttpParams, HttpResponse, HttpEvent }                           from '@angular/common/http';
 import { CustomHttpUrlEncodingCodec }                        from '../encoder';
 
-import { Observable }                                        from 'rxjs/Observable';
+import { Observable }                                        from 'rxjs';
 
 
 import { BASE_PATH, COLLECTION_FORMATS }                     from '../variables';
@@ -56,11 +55,11 @@ export class ReportsService {
 
 
     /**
-     * 
-     * 
-     * @param species 
-     * @param genomicDatasets 
-     * @param repDatasets 
+     *
+     *
+     * @param species
+     * @param genomicDatasets
+     * @param repDatasets
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -111,16 +110,16 @@ export class ReportsService {
     }
 
     /**
-     * 
-     * 
-     * @param reportName 
-     * @param format 
-     * @param species 
-     * @param genomicDatasets 
-     * @param genomicFilters 
-     * @param repDatasets 
-     * @param repFilters 
-     * @param params 
+     *
+     *
+     * @param reportName
+     * @param format
+     * @param species
+     * @param genomicDatasets
+     * @param genomicFilters
+     * @param repDatasets
+     * @param repFilters
+     * @param params
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -191,9 +190,9 @@ export class ReportsService {
     }
 
     /**
-     * 
-     * 
-     * @param jobId 
+     *
+     *
+     * @param jobId
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
