@@ -3,7 +3,7 @@
  * API for Ig Receptor gene data
  *
  * OpenAPI spec version: 1.0
- * 
+ *
  *
  * NOTE: This class is auto generated by the swagger code generator program.
  * https://github.com/swagger-api/swagger-codegen.git
@@ -12,11 +12,10 @@
 /* tslint:disable:no-unused-variable member-ordering */
 
 import { Inject, Injectable, Optional }                      from '@angular/core';
-import { HttpClient, HttpHeaders, HttpParams,
-         HttpResponse, HttpEvent }                           from '@angular/common/http';
+import { HttpClient, HttpHeaders, HttpParams, HttpResponse, HttpEvent }                           from '@angular/common/http';
 import { CustomHttpUrlEncodingCodec }                        from '../encoder';
 
-import { Observable }                                        from 'rxjs/Observable';
+import { Observable }                                        from 'rxjs';
 
 
 import { BASE_PATH, COLLECTION_FORMATS }                     from '../variables';
@@ -57,9 +56,9 @@ export class RepseqService {
 
     /**
      * Returns information on all samples
-     * 
-     * @param species 
-     * @param dataset 
+     *
+     * @param species
+     * @param dataset
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -104,8 +103,8 @@ export class RepseqService {
 
     /**
      * Return genotypes for all subjects of the specified species in the specified data type
-     * 
-     * @param species 
+     *
+     * @param species
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -146,8 +145,8 @@ export class RepseqService {
 
     /**
      * Returns the list of datasets available for the selected species
-     * 
-     * @param species 
+     *
+     * @param species
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -188,9 +187,9 @@ export class RepseqService {
 
     /**
      * Returns information and statistics on the dataset
-     * 
-     * @param species 
-     * @param dataset 
+     *
+     * @param species
+     * @param dataset
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -287,9 +286,9 @@ export class RepseqService {
 
     /**
      * Returns the inferred genotype (in MiAIRR format) of the specified sample
-     * 
-     * @param species 
-     * @param subjectName 
+     *
+     * @param species
+     * @param subjectName
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -334,7 +333,7 @@ export class RepseqService {
 
     /**
      * Returns the list all novel alleles across all datasets
-     * 
+     *
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -371,9 +370,9 @@ export class RepseqService {
 
     /**
      * Returns details of full-length novel alleles in a single dataset
-     * 
-     * @param species 
-     * @param dataset 
+     *
+     * @param species
+     * @param dataset
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -418,10 +417,10 @@ export class RepseqService {
 
     /**
      * Returns information on the selected sample
-     * 
-     * @param species 
-     * @param dataset 
-     * @param sample 
+     *
+     * @param species
+     * @param dataset
+     * @param sample
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -470,14 +469,14 @@ export class RepseqService {
 
     /**
      * Returns the list of samples in the selected dataset
-     * 
-     * @param species 
-     * @param dataset 
-     * @param pageNumber 
-     * @param pageSize 
-     * @param filter 
-     * @param sortBy 
-     * @param cols 
+     *
+     * @param species
+     * @param dataset
+     * @param pageNumber
+     * @param pageSize
+     * @param filter
+     * @param sortBy
+     * @param cols
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -545,14 +544,14 @@ export class RepseqService {
 
     /**
      * Returns the list of sequences in the selected datasets
-     * 
-     * @param species 
-     * @param dataset 
-     * @param pageNumber 
-     * @param pageSize 
-     * @param filter 
-     * @param sortBy 
-     * @param cols 
+     *
+     * @param species
+     * @param dataset
+     * @param pageNumber
+     * @param pageSize
+     * @param filter
+     * @param sortBy
+     * @param cols
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
@@ -620,7 +619,7 @@ export class RepseqService {
 
     /**
      * Returns the list of species for which information is held
-     * 
+     *
      * @param observe set whether or not to return the data Observable as the body, response or events. defaults to returning the body.
      * @param reportProgress flag to report request and response progress.
      */
