@@ -1,37 +1,46 @@
 /* tslint:disable:max-line-length */
 import { Component, Input, OnDestroy, OnInit, ViewChild, AfterViewInit, ViewEncapsulation, ElementRef } from '@angular/core';
-import { GenomicService } from '../../../../dist/digby-swagger-client';
+import { GenomicService } from 'projects/digby-swagger-client';
 import { GeneTableSelection } from '../../gene-table-selector/gene-table-selector.model';
 import { GeneTableSelectorService } from '../../gene-table-selector/gene-table-selector.service';
-import {NgbModal} from '@ng-bootstrap/ng-bootstrap';
-import {SeqModalComponent} from '../../seq-modal/seq-modal.component';
-import {GeneSequenceDataSource} from '../gene-sequence.datasource';
-import {MatPaginator} from '@angular/material/paginator';
-import {MatTable} from '@angular/material/table';
-import {columnInfo} from './gene-table-panel-cols';
-import {FilterMode} from '../../table/filter/filter-mode.enum';
-import {BehaviorSubject, fromEvent, Observable} from 'rxjs';
-import {IChoices} from '../../table/filter/ichoices';
-import {ColumnPredicate} from '../../table/filter/column-predicate';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
+import { SeqModalComponent } from '../../seq-modal/seq-modal.component';
+import { GeneSequenceDataSource } from '../gene-sequence.datasource';
+import { MatPaginator } from '@angular/material/paginator';
+import { MatTable, MatColumnDef, MatHeaderCellDef, MatHeaderCell, MatCellDef, MatCell, MatHeaderRowDef, MatHeaderRow, MatRowDef, MatRow } from '@angular/material/table';
+import { columnInfo } from './gene-table-panel-cols';
+import { FilterMode } from '../../table/filter/filter-mode.enum';
+import { BehaviorSubject, fromEvent, Observable } from 'rxjs';
+import { IChoices } from '../../table/filter/ichoices';
+import { ColumnPredicate } from '../../table/filter/column-predicate';
 import { GenGeneSelectedService } from '../gen-gene-selected.service'
-import {GenSampleSelectedService} from '../../gen-sample/gen-sample-selected.service';
-import { ResizeEvent } from 'angular-resizable-element';
-import { TableParamsStorageService } from '../../table/table-params-storage-service';
-import {debounceTime, distinctUntilChanged, filter, map} from 'rxjs/operators';
-import {NavigationEnd, Router} from '@angular/router';
-import {ReportRunService} from '../../reports/report-run.service';
-import {listsOfDictionariesEqual} from '../../shared/struct_utils';
+import { GenSampleSelectedService } from '../../gen-sample/gen-sample-selected.service';
+import { debounceTime, distinctUntilChanged, map } from 'rxjs/operators';
+import { Router } from '@angular/router';
+import { ReportRunService } from '../../reports/report-run.service';
+import { listsOfDictionariesEqual } from '../../shared/struct_utils';
+import { AsyncPipe } from '@angular/common';
+import { MatIcon } from '@angular/material/icon';
+import { ColumnSorterComponent } from '../../table/column-sorter/column-sorter.component';
+import { MatMenuTrigger, MatMenu, MatMenuContent } from '@angular/material/menu';
+import { FormsModule } from '@angular/forms';
+import { TableResizeDirective } from '../../table/table-resize.directive';
+import { ColumnResizeDirective } from '../../table/column-resize.directive';
+import { FilterComponent } from '../../table/filter/filter.component';
 
 
 @Component({
-  selector: 'app-gene-table-panel',
-  templateUrl: './gen-gene-table-panel.component.html',
-  styleUrls: ['./gen-gene-table-panel.component.css'],
-  encapsulation: ViewEncapsulation.None,   // needed for css styling on mat-menu-panel
-  providers: [TableParamsStorageService],
+    selector: 'app-gene-table-panel',
+    templateUrl: './gen-gene-table-panel.component.html',
+    styleUrls: ['./gen-gene-table-panel.component.css'],
+    encapsulation: ViewEncapsulation.None, // needed for css styling on mat-menu-panel
+    providers: [],
+    imports: [MatIcon, ColumnSorterComponent, MatMenuTrigger, FormsModule, MatPaginator, MatTable, TableResizeDirective, MatColumnDef, MatHeaderCellDef, MatHeaderCell, ColumnResizeDirective, FilterComponent, MatCellDef, MatCell, MatHeaderRowDef, MatHeaderRow, MatRowDef, MatRow, MatMenu, MatMenuContent, AsyncPipe]
 })
 
 export class GenGeneTablePanelComponent implements AfterViewInit, OnInit, OnDestroy {
+  // TODO: Skipped for migration because:
+  //  Your application code writes to the input. This prevents migration.
   @Input() selection: GeneTableSelection;
   @ViewChild(MatPaginator) paginator: MatPaginator;
   @ViewChild(MatTable) table: MatTable<string>;
@@ -53,7 +62,6 @@ export class GenGeneTablePanelComponent implements AfterViewInit, OnInit, OnDest
   selectedSampleIds = [];
   samplesSelected = false;
   isSelectedGenesChecked = false;
-  resizeEvents = new Map();
   clearSubject = new BehaviorSubject<null>(null);
   clear$ = this.clearSubject.asObservable();
   setFilterSubject = new BehaviorSubject<any>(null);
@@ -74,17 +82,14 @@ export class GenGeneTablePanelComponent implements AfterViewInit, OnInit, OnDest
               private geneTableService: GeneTableSelectorService,
               private modalService: NgbModal,
               private genGeneSelectedService: GenGeneSelectedService,
-              private tableParamsStorageService: TableParamsStorageService,
               private genSampleSelectedService: GenSampleSelectedService,
               private router: Router,
               private reportRunService: ReportRunService,
-              private el: ElementRef,
   ) {
   }
 
 
   ngOnInit() {
-    this.resizeEvents = this.tableParamsStorageService.loadSavedInfo(this.resizeEvents, 'gen-gene-table-widths');
     this.dataSource = new GeneSequenceDataSource(this.genomicService);
   }
 
@@ -121,8 +126,6 @@ export class GenGeneTablePanelComponent implements AfterViewInit, OnInit, OnDest
       this.loading$Subscription = this.dataSource.loading$.subscribe(
         loading => {
           if (!loading) {
-            this.applyResizes();
-
             if (this.redirectOnLoad) {
               const r = this.redirectOnLoad;
               this.redirectOnLoad = null;
@@ -149,13 +152,6 @@ export class GenGeneTablePanelComponent implements AfterViewInit, OnInit, OnDest
         }
       );
 
-      this.loading$Subscription = this.dataSource.loading$.subscribe(
-        loading => {
-          if (!loading) {
-            this.applyResizes();
-          }
-        }
-      );
 
       this.genSampleSelectedServiceSubscription = this.genSampleSelectedService.source.subscribe(
         selectedIds => {
@@ -177,11 +173,6 @@ export class GenGeneTablePanelComponent implements AfterViewInit, OnInit, OnDest
         }
       );
 
-      this.router.events.subscribe((val) => {
-        if (val instanceof NavigationEnd) {
-          this.applyResizes();
-        }
-      });
 
       this.loadSequences$Subscription = this.loadSequences$.pipe(debounceTime(500)).subscribe(() => {
         this.loadSequencesPage();
@@ -288,8 +279,6 @@ export class GenGeneTablePanelComponent implements AfterViewInit, OnInit, OnDest
     const deleted = difference(sLoaded, sDisplayed);
     this.filters = this.filters.filter((x) => !(deleted.has(x.field)));
     this.sorts = this.sorts.filter((x) => !(deleted.has(x.field)));
-
-    this.applyResizes();
   }
 
   loadSequencesPage() {
@@ -322,29 +311,6 @@ export class GenGeneTablePanelComponent implements AfterViewInit, OnInit, OnDest
     this.setFilterSubject.next({operator: { name: 'Includes', operands: 2, operator: 'like', prefix: '', postfix: '' }, op1: seq.name, op2: ''});
   }
 
-  onResizeEnd(event: ResizeEvent, columnName): void {
-    if (event.edges.right) {
-      const cssValue = event.rectangle.width + 'px';
-      this.updateColumnWidth(columnName, cssValue);
-      this.resizeEvents.set(columnName, cssValue);
-      this.tableParamsStorageService.saveInfo(this.resizeEvents, 'gen-gene-table-widths');
-    }
-  }
-
-  applyResizes(): void {
-    for (const [columnName, cssValue] of this.resizeEvents) {
-      this.updateColumnWidth(columnName, cssValue);
-    }
-  }
-
-  updateColumnWidth(columnName: string, cssValue: string) {
-    const columnElts = this.el.nativeElement.getElementsByClassName('mat-column-' + columnName);
-    // tslint:disable-next-line:prefer-for-of
-    for (let i = 0; i < columnElts.length; i++) {
-      const currentEl = columnElts[i] as HTMLDivElement;
-      currentEl.style.width = cssValue;
-    }
-  }
 
   sendReportRequest(report, format, params) {
     const datasets = this.selection.datasets;
@@ -384,4 +350,3 @@ function sampleIdsEqual(id1, id2) {
 
   return Object.keys(id1).every((v) => arraysEqual(id1[v], id2[v]));
 }
-
