@@ -2,36 +2,46 @@
 import { Component, Input, OnDestroy, OnInit, ViewChild, AfterViewInit, ViewEncapsulation, ElementRef } from '@angular/core';
 import { GeneTableSelection } from '../../gene-table-selector/gene-table-selector.model';
 import { GeneTableSelectorService } from '../../gene-table-selector/gene-table-selector.service';
-import {MatPaginator} from '@angular/material/paginator';
-import {MatTable} from '@angular/material/table';
-import {GenSampleDataSource} from '../gen-sample-data.source';
+import { MatPaginator } from '@angular/material/paginator';
+import { MatTable, MatColumnDef, MatHeaderCellDef, MatHeaderCell, MatCellDef, MatCell, MatHeaderRowDef, MatHeaderRow, MatRowDef, MatRow } from '@angular/material/table';
+import { GenSampleDataSource } from '../gen-sample-data.source';
 import { FilterMode } from '../../table/filter/filter-mode.enum';
 import { ColumnPredicate } from '../../table/filter/column-predicate';
 import { IChoices } from '../../table/filter/ichoices';
-import {BehaviorSubject, fromEvent, Observable, Subscription} from 'rxjs';
+import { BehaviorSubject, fromEvent, Observable, Subscription } from 'rxjs';
 import { columnInfo } from './gen-sample-panel-cols';
-import {NgbModal} from '@ng-bootstrap/ng-bootstrap';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { GenSampleInfoComponent } from '../gen-sample-info/gen-sample-info.component';
-import { GenomicService } from '../../../../dist/digby-swagger-client';
-import {GenGeneSelectedService} from '../../gen-gene-table/gen-gene-selected.service';
-import {GenSampleSelectedService} from '../gen-sample-selected.service';
-import {GenSampleFilterService} from '../gen-sample-filter.service';
-import { ResizeEvent } from 'angular-resizable-element';
-import {TableParamsStorageService} from '../../table/table-params-storage-service';
-import {ActivatedRoute, NavigationEnd, ParamMap, Router} from '@angular/router';
-import {debounceTime, distinctUntilChanged, filter, map} from 'rxjs/operators';
-import {ReportRunService} from '../../reports/report-run.service';
+import { GenomicService } from 'projects/digby-swagger-client';
+import { GenGeneSelectedService } from '../../gen-gene-table/gen-gene-selected.service';
+import { GenSampleSelectedService } from '../gen-sample-selected.service';
+import { GenSampleFilterService } from '../gen-sample-filter.service';
+import { ActivatedRoute, Router } from '@angular/router';
+import { debounceTime, distinctUntilChanged, map } from 'rxjs/operators';
+import { ReportRunService } from '../../reports/report-run.service';
+import { AsyncPipe } from '@angular/common';
+import { MatIcon } from '@angular/material/icon';
+import { ColumnSorterComponent } from '../../table/column-sorter/column-sorter.component';
+import { MatMenuTrigger, MatMenu, MatMenuContent } from '@angular/material/menu';
+import { FormsModule } from '@angular/forms';
+import { TableResizeDirective } from '../../table/table-resize.directive';
+import { ColumnResizeDirective } from '../../table/column-resize.directive';
+import { FilterComponent } from '../../table/filter/filter.component';
 
 
 @Component({
-  selector: 'app-gen-sample-panel',
-  templateUrl: './gen-sample-panel.component.html',
-  styleUrls: ['./gen-sample-panel.component.css'],
-  providers: [TableParamsStorageService],
-  encapsulation: ViewEncapsulation.None   // needed for css styling on mat-menu-panel
+    selector: 'app-gen-sample-panel',
+    templateUrl: './gen-sample-panel.component.html',
+    styleUrls: ['./gen-sample-panel.component.css'],
+    providers: [],
+    encapsulation: ViewEncapsulation.None // needed for css styling on mat-menu-panel
+    ,
+    imports: [MatIcon, ColumnSorterComponent, MatMenuTrigger, FormsModule, MatPaginator, MatTable, TableResizeDirective, MatColumnDef, MatHeaderCellDef, MatHeaderCell, ColumnResizeDirective, FilterComponent, MatCellDef, MatCell, MatHeaderRowDef, MatHeaderRow, MatRowDef, MatRow, MatMenu, MatMenuContent, AsyncPipe]
 })
 
 export class GenSamplePanelComponent implements AfterViewInit, OnInit, OnDestroy {
+  // TODO: Skipped for migration because:
+  //  Your application code writes to the input. This prevents migration.
   @Input() selection: GeneTableSelection;
   @ViewChild(MatPaginator) paginator: MatPaginator;
   @ViewChild(MatTable) table: MatTable<string>;
@@ -54,7 +64,6 @@ export class GenSamplePanelComponent implements AfterViewInit, OnInit, OnDestroy
   loading$Subscription = null;
   selectedSequenceNames: string[] = [];
   isSelectedSamplesChecked = false;
-  resizeEvents = new Map();
   clearSubject = new BehaviorSubject<null>(null);
   clear$ = this.clearSubject.asObservable();
   setFilterSubject = new BehaviorSubject<any>(null);
@@ -66,17 +75,11 @@ export class GenSamplePanelComponent implements AfterViewInit, OnInit, OnDestroy
               private genGeneSelectedService: GenGeneSelectedService,
               private genSampleSelectedService: GenSampleSelectedService,
               private genSampleFilterService: GenSampleFilterService,
-              private tableParamsStorageService: TableParamsStorageService,
-              private route: ActivatedRoute,
               private reportRunService: ReportRunService,
-              private router: Router,
-              private el: ElementRef,
               ) {
-
   }
 
   ngOnInit() {
-    this.resizeEvents = this.tableParamsStorageService.loadSavedInfo(this.resizeEvents, 'gen-sample-table-widths');
     this.dataSource = new GenSampleDataSource(this.genSampleService);
   }
 
@@ -96,9 +99,9 @@ export class GenSamplePanelComponent implements AfterViewInit, OnInit, OnDestroy
         .pipe(debounceTime(500)).subscribe(
           (sel: GeneTableSelection) => {
             this.selection = sel;
+            this.onSelectedSamplesChange();
             this.paginator.firstPage();
             this.table.renderRows();
-            this.loadSequencesPage();
           }
         );
 
@@ -108,14 +111,6 @@ export class GenSamplePanelComponent implements AfterViewInit, OnInit, OnDestroy
             this.genSampleSelectedService.selection.next({ids: choices.names_by_dataset});
           } else {
             this.genSampleSelectedService.selection.next({ids: []});
-          }
-        }
-      );
-
-      this.loading$Subscription = this.dataSource.loading$.subscribe(
-        loading => {
-          if (!loading) {
-            this.applyResizes();
           }
         }
       );
@@ -134,11 +129,6 @@ export class GenSamplePanelComponent implements AfterViewInit, OnInit, OnDestroy
 
         }
       );
-      this.router.events.subscribe((val) => {
-        if (val instanceof NavigationEnd) {
-          this.applyResizes();
-        }
-      });
 
       const searchOrder = {field: 'sample_name', predicates: [], sort: {field: 'sample_name', order: 'asc'}}
       this.applyFilter(searchOrder);
@@ -225,6 +215,7 @@ export class GenSamplePanelComponent implements AfterViewInit, OnInit, OnDestroy
   }
 
   updateColumnData(event: any) {
+    console.log("updateColumnData", event);
     const sLoaded = new Set(this.lastLoadedColumns);
     const sDisplayed = new Set(this.displayedColumns);
 
@@ -257,29 +248,6 @@ export class GenSamplePanelComponent implements AfterViewInit, OnInit, OnDestroy
     modalRef.componentInstance.dataset = sample.dataset;
   }
 
-  onResizeEnd(event: ResizeEvent, columnName): void {
-    if (event.edges.right) {
-      const cssValue = event.rectangle.width + 'px';
-      this.updateColumnWidth(columnName, cssValue);
-      this.resizeEvents.set(columnName, cssValue);
-      this.tableParamsStorageService.saveInfo(this.resizeEvents, 'rep-sample-table-widths');
-    }
-  }
-
-  applyResizes(): void {
-    for (const [columnName, cssValue] of this.resizeEvents) {
-      this.updateColumnWidth(columnName, cssValue);
-    }
-  }
-
-  updateColumnWidth(columnName: string, cssValue: string) {
-    const columnElts = this.el.nativeElement.getElementsByClassName('mat-column-' + columnName);
-    // tslint:disable-next-line:prefer-for-of
-    for (let i = 0; i < columnElts.length; i++) {
-      const currentEl = columnElts[i] as HTMLDivElement;
-      currentEl.style.width = cssValue;
-    }
-  }
 
   sendReportRequest(report, format, params) {
     let reportParams = {};
