@@ -1,12 +1,13 @@
-import {Input, Output, EventEmitter, Component} from '@angular/core';
+import {Output, EventEmitter, Component, input} from '@angular/core';
 import { ColumnPredicate } from './column-predicate';
 import { IChoices } from './ichoices';
 import { Observable } from 'rxjs';
 
 @Component({
-  template: ''
+    template: '',
+    standalone: false
 })export class FilterImplementation {
-  @Input() columnName: string;
-  @Input() choices$: Observable<IChoices>;
+  readonly columnName = input<string>(undefined);
+  readonly choices$ = input<Observable<IChoices>>(undefined);
   @Output() predicateEmitter = new EventEmitter<ColumnPredicate>();
 }
