@@ -1,17 +1,20 @@
 import {Component, OnInit, ViewChild, ViewEncapsulation} from '@angular/core';
 import {GeneTableSelection} from '../../gene-table-selector/gene-table-selector.model';
 import {GeneTableSelectorService} from '../../gene-table-selector/gene-table-selector.service';
-import { GenomicService } from '../../../../dist/digby-swagger-client';
-import { GoogleChartsModule } from 'angular-google-charts';
-import {MatTable} from '@angular/material/table';
+import { GenomicService } from 'projects/digby-swagger-client';
+import { GoogleChartsModule, ChartType } from 'angular-google-charts';
+import { MatTable, MatColumnDef, MatHeaderCellDef, MatHeaderCell, MatCellDef, MatCell, MatHeaderRowDef, MatHeaderRow, MatRowDef, MatRow } from '@angular/material/table';
 import {debounceTime} from "rxjs/operators";
 
 
+
 @Component({
-  selector: 'app-gen-explore-data-panel',
-  templateUrl: './gen-explore-data-panel.component.html',
-  styleUrls: ['./gen-explore-data-panel.component.css'],
-  encapsulation: ViewEncapsulation.None   // needed for css styling on mat-menu-panel
+    selector: 'app-gen-explore-data-panel',
+    templateUrl: './gen-explore-data-panel.component.html',
+    styleUrls: ['./gen-explore-data-panel.component.css'],
+    encapsulation: ViewEncapsulation.None // needed for css styling on mat-menu-panel
+    ,
+    imports: [MatTable, MatColumnDef, MatHeaderCellDef, MatHeaderCell, MatCellDef, MatCell, MatHeaderRowDef, MatHeaderRow, MatRowDef, MatRow, GoogleChartsModule]
 })
 
 export class GenExploreDataPanelComponent implements OnInit {
@@ -22,6 +25,8 @@ export class GenExploreDataPanelComponent implements OnInit {
   datasetInfo = null;
   loading = false;
   error = null;
+  barChartType = ChartType.BarChart;
+  pieChartType = ChartType.PieChart;
   @ViewChild('datasetInfoTable') datasetInfoTable: MatTable<any>;
 
   constructor(
