@@ -1,9 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 
 @Component({
-  selector: 'app-licensing',
-  templateUrl: './licensing.component.html',
-  styleUrls: ['./licensing.component.scss']
+    selector: 'app-licensing',
+    templateUrl: './licensing.component.html',
+    styleUrls: ['./licensing.component.scss']
 })
 export class LicensingComponent implements OnInit {
 
