@@ -1,9 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 
 @Component({
-  selector: 'app-quick-ref',
-  templateUrl: './quick-ref.component.html',
-  styleUrls: ['./quick-ref.component.css']
+    selector: 'app-quick-ref',
+    templateUrl: './quick-ref.component.html',
+    styleUrls: ['./quick-ref.component.css']
 })
 export class QuickRefComponent implements OnInit {
 
