@@ -6,9 +6,9 @@ import {debounceTime, distinctUntilChanged, filter, map} from 'rxjs/operators';
 
 
 @Component({
-  selector: 'app-refbook-panel',
-  templateUrl: './refbook-panel.component.html',
-  styleUrls: ['./refbook-panel.component.css']
+    selector: 'app-refbook-panel',
+    templateUrl: './refbook-panel.component.html',
+    styleUrls: ['./refbook-panel.component.css']
 })
 
 export class RefbookPanelComponent implements OnInit, AfterViewInit {
