@@ -1,8 +1,9 @@
-import {Component, Input, OnInit, ViewEncapsulation} from '@angular/core';
-import { FormGroup, FormControl, Validators } from '@angular/forms';
+import {Component, OnInit, ViewEncapsulation, Input} from '@angular/core';
+import { FormGroup, FormControl, Validators, FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { NgbActiveModal, NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import {Subject} from 'rxjs';
-import {IDropdownSettings} from 'ng-multiselect-dropdown';
+import { IDropdownSettings, NgMultiSelectDropDownModule } from 'ng-multiselect-dropdown';
+
 
 export interface SampleCoverage {
   genomic: string[];
@@ -12,10 +13,12 @@ export interface SampleCoverage {
 }
 
 @Component({
-  selector: 'app-report-params-dialog',
-  templateUrl: './report-params-dialog.component.html',
-  styleUrls: ['./report-params-dialog.component.css'],
-  encapsulation: ViewEncapsulation.None   // needed for css styling
+    selector: 'app-report-params-dialog',
+    templateUrl: './report-params-dialog.component.html',
+    styleUrls: ['./report-params-dialog.component.css'],
+    encapsulation: ViewEncapsulation.None // needed for css styling
+    ,
+    imports: [FormsModule, ReactiveFormsModule, NgMultiSelectDropDownModule]
 })
 export class ReportParamsDialogComponent implements OnInit {
   @Input() report;
@@ -46,7 +49,8 @@ export class ReportParamsDialogComponent implements OnInit {
       this.add_fields(t, group);
     });
 
-    if (this.report.filter_params) {
+    const report = this.report;
+    if (report.filter_params) {
       for (let key in this.filterParams) {
         this.add_fields(this.filterParams[key], group);
       }
@@ -55,25 +59,26 @@ export class ReportParamsDialogComponent implements OnInit {
     this.myFormGroup = new FormGroup(group);
 
     this.scopeText = 'The report will be run on ';
-    if (this.report.scope.includes("gen_sample")) {
+    const sampleCoverage = this.sampleCoverage;
+    if (report.scope.includes("gen_sample")) {
       if (this.sampleCoverage.genomic.length > 1) {
         this.scopeText += this.sampleCoverage.genomicScope + ' samples from ' + this.sampleCoverage.genomic.join(', ') + ' genomic datasets';
-      } else if (this.sampleCoverage.genomic.length === 1) {
-        this.scopeText += this.sampleCoverage.genomicScope + ' samples from the ' + this.sampleCoverage.genomic[0] + ' genomic dataset';
+      } else if (sampleCoverage.genomic.length === 1) {
+        this.scopeText += sampleCoverage.genomicScope + ' samples from the ' + sampleCoverage.genomic[0] + ' genomic dataset';
       }
     }
 
-    if (this.report.scope.includes("gen_sample") &&
-      this.report.scope.includes("rep_sample") &&
-      this.sampleCoverage.genomic.length > 0 && this.sampleCoverage.AIRRSeq.length > 0) {
+    if (report.scope.includes("gen_sample") &&
+      report.scope.includes("rep_sample") &&
+      sampleCoverage.genomic.length > 0 && sampleCoverage.AIRRSeq.length > 0) {
       this.scopeText += ' and ';
     }
 
-    if (this.report.scope.includes("rep_sample")) {
-      if (this.sampleCoverage.AIRRSeq.length > 1) {
-        this.scopeText += this.sampleCoverage.AIRRSeqScope + ' samples from ' + this.sampleCoverage.AIRRSeq.join(', ') + ' AIRR-Seq datasets';
-      } else if (this.sampleCoverage.AIRRSeq.length === 1) {
-        this.scopeText += this.sampleCoverage.AIRRSeqScope + ' samples from the ' + this.sampleCoverage.AIRRSeq[0] + ' AIRR-Seq dataset';
+    if (report.scope.includes("rep_sample")) {
+      if (sampleCoverage.AIRRSeq.length > 1) {
+        this.scopeText += sampleCoverage.AIRRSeqScope + ' samples from ' + sampleCoverage.AIRRSeq.join(', ') + ' AIRR-Seq datasets';
+      } else if (sampleCoverage.AIRRSeq.length === 1) {
+        this.scopeText += sampleCoverage.AIRRSeqScope + ' samples from the ' + sampleCoverage.AIRRSeq[0] + ' AIRR-Seq dataset';
       }
     }
   }
