@@ -1,9 +1,7 @@
 import { Pipe, PipeTransform } from '@angular/core';
 import {DomSanitizer, SafeHtml, SafeResourceUrl, SafeScript, SafeStyle, SafeUrl} from '@angular/platform-browser';
 
-@Pipe({
-  name: 'safe'
-})
+@Pipe({ name: 'safe' })
 export class SafePipe implements PipeTransform {
 
   constructor(protected sanitizer: DomSanitizer) {}
