@@ -1,8 +1,6 @@
 import {Directive, ElementRef, HostBinding, HostListener} from '@angular/core';
 
-@Directive({
-  selector: '[appDropdown]'
-})
+@Directive({ selector: '[appDropdown]' })
 export class DropdownDirective {
   @HostBinding('class.open') isOpen = false;
   @HostListener('document:click', ['$event']) toggleOpen(event: Event) {
